wip: tidy rust
