+++ rust/.gitignore
@@ -195,3 +195,5 @@ Cargo.lock
 # !.yarn/versions
 
 *.node
+
+perf.data*
